Add exFAT undelete and unify FAT/exFAT recovery
Extends the `undelete` command to recover deleted files from exFAT volumes,
the default filesystem on SD/SDXC cards over 32 GB and most modern cameras.

- New `exfat` module: parse the boot sector, walk directory entry sets, and
  recover deleted files. exFAT deletion only clears the InUse bit, so the full
  name and metadata survive (no lost first character as in FAT).
- Handle contiguous files via the NoFatChain flag and fragmented files by
  following the FAT chain, with a contiguous fallback when the chain was freed.
- New `recover` module: shared `RecoverStats` plus a `Volume` dispatcher that
  auto-detects FAT vs exFAT (bare volume or MBR partitions, including type 0x07
  disambiguated from NTFS by signature).
- CLI `undelete` now auto-detects and recovers from either filesystem.
- Integration tests build synthetic exFAT images and verify recovery of a
  root-level file (with a space in its long name) and a file inside a live
  subdirectory, byte-for-byte.
- README/lib docs updated to cover both FAT and exFAT.

+++ src/cli.rs
@@ -20,11 +20,12 @@ pub enum Command {
     /// Works even after a format or partition loss, but cannot restore
     /// original filenames.
     Scan(ScanArgs),
-    /// Recover deleted files from a FAT filesystem, keeping their original
-    /// names, paths, and sizes.
+    /// Recover deleted files from a FAT or exFAT filesystem, keeping their
+    /// original names, paths, and sizes.
     ///
     /// More accurate than carving when the filesystem metadata is intact (e.g.
-    /// a file was just deleted), but requires a readable FAT12/16/32 volume.
+    /// a file was just deleted), but requires a readable FAT12/16/32 or exFAT
+    /// volume.
     Undelete(UndeleteArgs),
     /// List the file types this build can recover.
     ListTypes,
@@ -41,14 +42,14 @@ pub struct UndeleteArgs {
     #[arg(short, long, value_name = "DIR", default_value = "recovered")]
     pub output: PathBuf,
 
-    /// Byte offset of the FAT volume within the source. By default the source
-    /// is auto-detected (bare volume or MBR partition table).
+    /// Byte offset of the volume within the source. By default the source is
+    /// auto-detected (bare FAT/exFAT volume or MBR partition table).
     #[arg(long, value_name = "BYTES")]
     pub offset: Option<u64>,
 
     /// Ignore deleted files smaller than this many bytes.
     #[arg(long, value_name = "BYTES", default_value_t = 0)]
-    pub min_size: u32,
+    pub min_size: u64,
 }
 
 #[derive(Parser)]
